requirements.txt
# FANUC CNC Monitor - Requirements

# Core dependencies
pyyaml>=6.0.1
asyncio-mqtt>=0.16.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
